requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
pymysql==1.1.1
cryptography==42.0.8
python-dotenv==1.0.1
jinja2==3.1.4
python-multipart==0.0.12
reportlab==4.2.2
aiosmtplib==3.0.1
email-validator==2.2.0
